refactor + reorganize in modules. Closes LOT-95.
