+++ backend/metering_billing/management/commands/execute_calculate_invoice.py
@@ -1,8 +1,6 @@
 import time
 
 from django.core.management.base import BaseCommand
-from django.db import connections
-from django.db.utils import OperationalError
 from metering_billing.tasks import calculate_invoice
