+++ backend/metering_billing/management/commands/initadmin.py
@@ -1,11 +1,8 @@
 import os
 
-from django.core.management import call_command
 from django.core.management.base import BaseCommand
-from django.db import connection, connections
-from django.db.utils import OperationalError
 from dotenv import load_dotenv
-from metering_billing.models import APIToken, Organization, User
+from metering_billing.models import Organization, User
 
 load_dotenv()
