+++ backend/metering_billing/management/commands/setup_tasks.py
@@ -1,9 +1,6 @@
-import os
-
 from django.core.management.base import BaseCommand
 from django_celery_beat.models import CrontabSchedule, IntervalSchedule, PeriodicTask
 from dotenv import load_dotenv
-from metering_billing.models import APIToken, Organization, User
 
 load_dotenv()
