+++ backend/metering_billing/management/commands/up_test_data.py
@@ -1,8 +1,6 @@
 import datetime
 import itertools
-import os
 import random
-import time
 import uuid
 from datetime import timezone
