+++ backend/metering_billing/permissions.py
@@ -1,7 +1,4 @@
 from django.utils.translation import gettext_lazy as _
-from drf_spectacular.extensions import OpenApiAuthenticationExtension
-from knox.auth import TokenAuthentication
-from knox.models import AuthToken
 from metering_billing.models import APIToken
 from rest_framework_api_key.permissions import BaseHasAPIKey
 
@@ -19,19 +16,3 @@ class HasUserAPIKey(BaseHasAPIKey):
             else:
                 raise KeyError("No API key found in request")
 
-
-# class KnoxTokenScheme(TokenAuthentication, OpenApiAuthenticationExtension):
-#     # target_class = 'knox.auth.TokenAuthentication'
-#     # name = 'knoxTokenAuth'
-#     # match_subclasses = True
-#     # priority = 1
-
-#     def get_security_definition(self, auto_schema):
-#         return {
-#             'type': 'apiKey',
-#             'in': 'header',
-#             'name': 'Authorization',
-#             'description': _(
-#                 'Token-based authentication with required prefix "%s"'
-#             ) % "Token"
-#         }
