+++ backend/metering_billing/utils/utils.py
@@ -2,94 +2,10 @@ import collections
 import datetime
 from datetime import timezone
 from decimal import ROUND_DOWN, ROUND_UP, Decimal
-from enum import Enum
 
 from dateutil.relativedelta import relativedelta
-from django.db import models
 from django.utils.translation import gettext_lazy as _
-from model_utils import Choices
-
-
-class INVOICE_STATUS(models.TextChoices):
-    DRAFT = ("draft", _("Draft"))
-    PAID = ("paid", _("Paid"))
-    UNPAID = ("unpaid", _("Unpaid"))
-
-
-class PAYMENT_PLANS(models.TextChoices):
-    SELF_HOSTED_FREE = ("self_hosted_free", _("Self-Hosted Free"))
-    CLOUD = ("cloud", _("Cloud"))
-    SELF_HOSTED_ENTERPRISE = ("self_hosted_enterprise", _("Self-Hosted Enterprise"))
-
-
-class METRIC_AGGREGATION(models.TextChoices):
-    COUNT = ("count", _("Count"))
-    SUM = ("sum", _("Sum"))
-    MAX = ("max", _("Max"))
-    MIN = ("min", _("Min"))
-    UNIQUE = ("unique", _("Unique"))
-    LATEST = ("latest", _("Latest"))
-    AVERAGE = ("average", _("Average"))
-
-
-class PAYMENT_PROVIDERS(models.TextChoices):
-    STRIPE = ("stripe", _("Stripe"))
-
-
-class METRIC_TYPE(models.TextChoices):
-    AGGREGATION = ("aggregation", _("Aggregatable"))
-    STATEFUL = ("stateful", _("State Logging"))
-
-
-class PLAN_INTERVAL(models.TextChoices):
-    WEEK = ("week", _("Week"))
-    MONTH = ("month", _("Month"))
-    YEAR = ("year", _("Year"))
-
-
-class REVENUE_CALC_GRANULARITY(models.TextChoices):
-    DAILY = ("day", _("Daily"))
-    TOTAL = ("total", _("Total"))
-
-
-class NUMERIC_FILTER_OPERATORS(models.TextChoices):
-    GTE = ("gte", _("Greater than or equal to"))
-    GT = ("gt", _("Greater than"))
-    EQ = ("eq", _("Equal to"))
-    LT = ("lt", _("Less than"))
-    LTE = ("lte", _("Less than or equal to"))
-
-
-class CATEGORICAL_FILTER_OPERATORS(models.TextChoices):
-    ISIN = ("isin", _("Is in"))
-    ISNOTIN = ("isnotin", _("Is not in"))
-
-
-class SUBSCRIPTION_STATUS(models.TextChoices):
-    ACTIVE = ("active", _("Active"))
-    ENDED = ("ended", _("Ended"))
-    NOT_STARTED = ("not_started", _("Not Started"))
-    CANCELED = ("canceled", _("Canceled"))
-
-
-class PLAN_STATUS(models.TextChoices):
-    ACTIVE = ("active", _("Active"))
-    INACTIVE = ("inactive", _("Inactive"))
-    EXPERIMENTAL = ("experimental", _("Experimental"))
-
-
-class BACKTEST_KPI(models.TextChoices):
-    TOTAL_REVENUE = ("total_revenue", _("Total Revenue"))
-
-
-class BACKTEST_STATUS(models.TextChoices):
-    RUNNING = ("running", _("Running"))
-    COMPLETED = ("completed", _("Completed"))
-
-
-class PRODUCT_STATUS(models.TextChoices):
-    ACTIVE = ("active", _("Active"))
-    DEPRECATED = ("deprecated", _("Deprecated"))
+from metering_billing.utils.enums import REVENUE_CALC_GRANULARITY
 
 
 def convert_to_decimal(value):
@@ -215,3 +131,8 @@ def periods_bwn_twodates(granularity, start_date, end_date):
 
 def now_plus_day():
     return datetime.datetime.now(datetime.timezone.utc) + relativedelta(days=+1)
+
+def dates_bwn_twodates(start_date, end_date):
+    days_btwn = (end_date - start_date).days
+    for n in range(days_btwn + 1):
+        yield start_date + relativedelta(days=n)
