+++ backend/metering_billing/webhooks.py
@@ -1,5 +1,7 @@
-import requests
 import json
+
+import requests
+
 from .models import Alert
